fix(lib): move up dir to prevent do from trying to build it
