+++ go.work
@@ -1,0 +1,8 @@
+go 1.19
+
+use (
+	.
+	./packages/website/mail
+	./packages/website/partner
+	./packages/website/upload
+)
