--- packages/go.work
@@ -1,9 +1,0 @@
-go 1.19
-
-use (
-	.
-	./website/mail
-	./website/partner
-	./website/pitching
-	./website/upload
-)
