+++ project.yml
@@ -38,3 +38,17 @@ packages:
         limits:
           memory: 1024
           timeout: 10000
+      - name: partner
+        binary: false
+        main: ""
+        runtime: go:1.20
+        web: true
+        webSecure: false
+        environment:
+          REGION: "${REGION}"
+          BUCKET: "${BUCKET}"
+          KEY: "${KEY}"
+          SECRET: "${SECRET}"
+        limits:
+          memory: 1024
+          timeout: 10000
